fix: change int to long

+++ 0x15-file_io/100-elf_header.c
@@ -64,7 +64,7 @@ void print_elf_header_info(const Elf64_Ehdr *header)
 	printf("  Type:                              %s\n",
 		   header->e_type == ET_EXEC ? "EXEC (Executable file)" : "Unknown type");
 	printf("  Entry point address:               0x%lx\n",
-		   (unsigned int) header->e_entry);
+		   (unsigned long) header->e_entry);
 }
 
 /**
